แก้ check box หน้า edit ฝั่งร้าน

+++ pixupfoodtest/restaurant-setting/edit-foodbox.php
@@ -1,18 +1,22 @@
 <?php
+
 session_start();
 include '../dbconn.php';
 $resid = $_SESSION["restdata"]["id"];
-$foodbox = $_POST["foodbox"];
+$foodbox = @$_POST["foodbox"];
 $foodbox_id = "";
 
-$con->query("DELETE FROM `mapping_food_box` WHERE restaurant_id = '$resid' ");
+if (isset($foodbox)) {
+
+    $con->query("DELETE FROM `mapping_food_box` WHERE restaurant_id = '$resid' ");
 
 
-foreach ($foodbox as $fb) {
-    $foodbox_id = $fb;
+    foreach ($foodbox as $fb) {
+        $foodbox_id = $fb;
 
-    $con->query("INSERT INTO `mapping_food_box`(`restaurant_id`, `food_box_id`)"
-            . " VALUES ('$resid','$foodbox_id')");
+        $con->query("INSERT INTO `mapping_food_box`(`restaurant_id`, `food_box_id`)"
+                . " VALUES ('$resid','$foodbox_id')");
+    }
 }
 
 if ($con->error == "") {
@@ -23,7 +27,7 @@ if ($con->error == "") {
     <?php
 
 } else {
-    
+
     echo $con->error;
 }
