+++ pixupfoodtest/restaurant-setting/edit-paymenttype.php
@@ -1,4 +1,5 @@
 <?php
+
 session_start();
 include '../dbconn.php';
 $resid = $_SESSION["restdata"]["id"];
@@ -6,13 +7,16 @@ $paytype = @$_POST["paymentData"];
 $payid = "";
 
 //print_r($_POST);
- $con->query("DELETE FROM `mapping_payment_type` WHERE restaurant_id = '$resid' ");
-  
-foreach ($paytype as $fb) {
-    $payid = $fb;
-    
-    $con->query("INSERT INTO `mapping_payment_type`(`restaurant_id`, `payment_type_id`)"
-            . " VALUES ('$resid','$payid')");
+
+
+if (isset($paytype)) {
+    $con->query("DELETE FROM `mapping_payment_type` WHERE restaurant_id = '$resid' ");
+    foreach ($paytype as $fb) {
+        $payid = $fb;
+
+        $con->query("INSERT INTO `mapping_payment_type`(`restaurant_id`, `payment_type_id`)"
+                . " VALUES ('$resid','$payid')");
+    }
 }
 if ($con->error == "") {
     ?>
@@ -23,7 +27,6 @@ if ($con->error == "") {
 
 } else {
     echo $con->error;
-
 }
